Refactoring successful
After a Domestika Flutter course, I have made a few attempts to refactor the code.
However, the code base is too large (500 words), so this will take some time.

From now on, widgets and functions will be in a separate dart files for clean code.
Progress: Change in data storage, backend, and additional custom keyboard widget (not in that order)