Create custom LED control widget/button
Moving LED widget as a custom class/widget - lots of work transitioning parameters.
But it was worth it and the code is much cleaner now.
New change: every button/widget has a separate SSH client now.

Normally, there might be a default limit for SSH clients of 10, this can however be changed.

Could just be me, but this new SSH implementation seems to reduce latency.

Anyway, after 4.5 hours of debug, gotta sleep.