UpdateRepo new implementation
Previously, updating repo is done by the appbar, leading to re-updating every time a button is added/removed & flutter development app is reloaded.
This is annoying for both user and developer.
By implementing this into a pop up (dialog), these problems are solved.
But sometimes this causes the space to close randomly... to be investigated further